refactored the main module. Also added a react component which handles all react unhandled exceptions
