+++ src/has_stack.ts
@@ -1,0 +1,12 @@
+import { NotifiableError } from "./index";
+
+export function hasStack(error: NotifiableError) {
+  return (
+    !!error &&
+    (!!error.stack || !!error.stacktrace || !!error["opera#sourceloc"]) &&
+    typeof (error.stack || error.stacktrace || error["opera#sourceloc"]) ===
+      "string" &&
+    error.stack !== `${error.name}: ${error.message}`
+  );
+}
+
